requirements.txt
azureml-sdk==1.0.39
scikit-learn==0.21.1
jupyter==1.0.0
matplotlib==3.1.0
joblib==0.13.2
scikit-plot==0.3.7
azureml-widgets==1.0.39
azureml-train-automl==1.0.39.1
